performance-testing/requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==36.0.2
idna==3.6
mnemonic==0.20
multiversx-sdk-core==0.8.0
multiversx-sdk-network-providers==0.13.0
multiversx-sdk-wallet==0.9.1
protobuf==3.20.1
pycparser==2.22
pycryptodomex==3.16.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
